Fix Ctrl+C exit delay and improve display update behaviour.

+++ src/PowerPing/PingResults.cs
@@ -50,7 +50,7 @@ namespace PowerPing
 
         // Local variables
         private readonly Stopwatch operationTimer = new Stopwatch();
-        private ulong sum  = 0; // Sum off all reply times
+        private double sum = 0; // Sum of all reply times
 
         public PingResults()
         {
@@ -94,8 +94,8 @@ namespace PowerPing
 
             try {
                 // Work out average
-                sum += (ulong)time;
-                AvgTime = (double)sum / Received; // Avg = Total / Count
+                sum += time;
+                AvgTime = sum / Received; // Avg = Total / Count
             } catch (OverflowException) {
                 HasOverflowed = true;
             }
